backend/requirements.txt
ag-ui-protocol==0.1.10
altair==5.5.0
annotated-doc==0.0.3
annotated-types==0.7.0
anthropic==0.72.0
anyio==4.11.0
argcomplete==3.6.3
asttokens==3.0.0
attrs==25.4.0
authlib==1.6.5
backcall==0.2.0
beartype==0.22.5
beautifulsoup4==4.14.2
bleach==6.3.0
blinker==1.9.0
boto3==1.40.69
botocore==1.40.69
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cohere==5.20.0
colorama==0.4.6
cryptography==46.0.3
cyclopts==4.2.1
datetime==5.5
decorator==5.2.1
defusedxml==0.7.1
diskcache==5.6.3
distro==1.9.0
dnspython==2.8.0
docopt==0.6.2
docstring-parser==0.17.0
docutils==0.22.3
email-validator==2.3.0
eval-type-backport==0.2.2
exceptiongroup==1.3.0
executing==2.2.1
fastapi==0.121.0
fastapi-cli==0.0.14
fastapi-cloud-cli==0.3.1
fastavro==1.12.1
fastjsonschema==2.21.2
fastmcp==2.13.0.2
filelock==3.20.0
filetype==1.2.0
fsspec==2025.10.0
genai-prices==0.0.37
gitdb==4.0.12
gitpython==3.1.45
google-ai-generativelanguage==0.9.0
google-api-core==2.28.1
google-auth==2.43.0
google-genai==1.49.0
googleapis-common-protos==1.72.0
greenlet==3.2.4
griffe==1.14.0
groq==0.33.0
grpcio==1.76.0
grpcio-status==1.76.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==1.1.2
idna==3.11
importlib-metadata==8.7.0
invoke==2.2.1
ipython==8.12.3
jaraco-classes==3.4.0
jaraco-context==6.0.1
jaraco-functools==4.3.0
jedi==0.19.2
jinja2==3.1.6
jiter==0.11.1
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-path==0.3.4
jsonschema-specifications==2025.9.1
jupyter-client==8.6.3
jupyter-core==5.9.1
jupyterlab-pygments==0.3.0
keyring==25.6.0
langchain==1.0.5
langchain-core==1.0.4
langchain-google-genai==3.0.1
langgraph==1.0.2
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.2
langgraph-sdk==0.2.9
langsmith==0.4.41
logfire==4.14.2
logfire-api==4.14.2
markdown-it-py==4.0.0
markupsafe==3.0.3
matplotlib-inline==0.2.1
mcp==1.21.0
mdurl==0.1.2
mistralai==1.9.11
mistune==3.1.4
more-itertools==10.8.0
narwhals==2.10.2
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nexus-rpc==1.1.0
numpy==2.3.4
openai==2.7.1
openapi-pydantic==0.5.1
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-http==1.38.0
opentelemetry-instrumentation==0.59b0
opentelemetry-instrumentation-httpx==0.59b0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
opentelemetry-util-http==0.59b0
orjson==3.11.4
ormsgpack==1.12.0
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.5
pathable==0.4.4
pathvalidate==3.3.1
pickleshare==0.7.5
pillow==12.0.0
pip==24.2
pipreqs==0.5.0
platformdirs==4.5.0
playwright==1.55.0
prompt-toolkit==3.0.52
proto-plus==1.26.1
protobuf==6.33.0
pure-eval==0.2.3
py-key-value-aio==0.2.8
py-key-value-shared==0.2.8
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.23
pydantic==2.12.4
pydantic-ai==1.12.0
pydantic-ai-slim==1.12.0
pydantic-core==2.41.5
pydantic-evals==1.12.0
pydantic-graph==1.12.0
pydantic-settings==2.11.0
pydeck==0.9.1
pyee==13.0.0
pygments==2.19.2
pyjwt==2.10.1
pyperclip==1.11.0
pysocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
pytz==2025.2
pywin32==311
pywin32-ctypes==0.2.3
pyyaml==6.0.3
pyzmq==27.1.0
referencing==0.36.2
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.2.0
rich-rst==1.3.2
rich-toolkit==0.15.1
rignore==0.7.6
rpds-py==0.28.0
rsa==4.9.1
s3transfer==0.14.0
selenium==4.38.0
sentry-sdk==2.43.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
sse-starlette==3.0.3
stack-data==0.6.3
starlette==0.49.3
streamlit==1.51.0
temporalio==1.18.2
tenacity==9.1.2
tinycss2==1.4.0
tokenizers==0.22.1
toml==0.10.2
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
trio==0.32.0
trio-websocket==0.12.2
typer==0.20.0
typer-slim==0.20.0
types-protobuf==6.32.1.20251105
types-requests==2.32.4.20250913
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
watchdog==6.0.0
watchfiles==1.1.1
wcwidth==0.2.14
webencodings==0.5.1
websocket-client==1.9.0
websockets==15.0.1
wrapt==1.17.3
wsproto==1.2.0
xxhash==3.6.0
yarg==0.1.9
zipp==3.23.0
zope-interface==8.0.1
zstandard==0.25.0
